Error Handling

Websockets are inherently asynchronous, so error handling can be inflexible.

SimpleWebsockets.jl offers the user event hooks to register callbacks and handle errors flexibly.

SEE: Server Events, Client Events, Connection Events

If no callbacks are registered, hardcoded @info, @warn and @error calls will provide logging feedback, but the parent process will remain unaffected by exceptions.

WebsocketError

SimpleWebsockets.WebsocketError — Type

abstract type WebsocketError <: Exception

WebsocketError child error types have the following fields:

msg::String

log::Function

The log() function will internally call: @error msg exception = (err, trace)

Where trace is the backtrace of the exception origin.

WebsocketError types

SimpleWebsockets.ConnectError — Type

struct ConnectError <: WebsocketError

An exception originated while trying to start a server or connect to a server

SimpleWebsockets.CallbackError — Type

struct CallbackError <: WebsocketError

An exception originated in a user provided callback function

SimpleWebsockets.FrameError — Type

struct FrameError <: WebsocketError

An exception originated while parsing a websocket data frame

Convenience methods

SimpleWebsockets.logWSerror — Function

logWSerror(err::Union{Exception, WebsocketError})

A convenience method to lift errors into a scope and log them.

Example

using SimpleWebsockets

server = WebsocketServer()
ended = Condition()

listen(server, :client) do client end
listen(server, :connectError) do err::WebsocketError
    notify(ended, err)
end

@async serve(server, 8080, "notahost")

reason = wait(ended)
reason isa Exception && logWSerror(reason)

SimpleWebsockets.throwWSerror — Function

throwWSerror(err::Union{Exception, WebsocketError})

A convenience method to lift fatal errors into a scope and throw them.

Example

using SimpleWebsockets

server = WebsocketServer()
ended = Condition()

listen(server, :client) do client end
listen(server, :connectError) do err::WebsocketError
    notify(ended, err)
end

@async serve(server, 8080, "notahost")

reason = wait(ended)
reason isa Exception && throwWSerror(reason)